add dynamic routing
