+++ components/EpisodeListItem/index.js
@@ -1,3 +1,4 @@
+import Link from "next/link";
 import styled from "styled-components";
 
 const StyledListItem = styled.li`
